Fix flakiness in using-dev-build (#611)
* Fix flakiness in using-dev-build

This fixes (most of) the flakiness in `using-dev-build`. There were
several issues:

* ChromeDriver was out of date and couldn't be updated; fixed by
  https://github.com/giggio/node-chromedriver/pull/358 and by then
  updating to work well with Chrome 100
* Tests sometimes started before proxy was up; we now wait for the proxy
  to be up before we start the tests
* Some tests timed out waiting for an element to appear; global timeout
  was changed from 5s to 10s.

* Update selenium standalone config

+++ demos/using-dev-build/test.js
@@ -60,48 +60,63 @@ const proxy = child_process.spawn("proxy", [
   `${canister_id}:${II_DAPP_PORT}`,
 ]);
 
+let testsStarted = false;
+let proxyOutput = "";
+
 proxy.stdout.on("data", (data) => {
   console.log(`proxy: ${data}`);
+
+  if(!testsStarted) {
+      // Wait for proxy to output "Proxy created" before we start the tests, otherwise the
+      // tests may start before the proxy is running.
+      proxyOutput = `${proxyOutput}${data}`;
+      if(proxyOutput.includes("Proxy created")) {
+
+          console.log("Starting tests");
+          testsStarted = true;
+
+          // TODO this should be a flag, and we should have
+          // npm run test # run the tests
+          // npm run start # start the proxy
+          if (true) {
+              /*
+               * Start the tests
+               */
+              const wdio = child_process.spawn("npm", ["run", "wdio"], {
+                  env: { ...process.env, II_DAPP_URL: II_DAPP_URL },
+              });
+
+              wdio.stdout.on("data", (data) => {
+                  console.log(`wdio: ${data}`);
+              });
+
+              wdio.stderr.on("data", (data) => {
+                  console.log(`wdio: ${data}`);
+              });
+
+              wdio.on('exit', (code, signal) => {
+                  console.log("Killing proxy");
+                  proxy.kill();
+                  if (code > 0) {
+                      // if code is set and non-zero, bubble up error from wdio tests
+                      throw new Error(`End-to-end tests returned with ${code}`);
+                  } else if (signal) {
+                      // otherwise, if the child was killed externally
+                      throw new Error(`End-to-end tests killed with signal ${signal}`);
+                  } else {
+                      console.log('End-to-end tests finished successfully');
+                  }
+              })
+          }
+
+      }
+  }
 });
 
 proxy.stderr.on("data", (data) => {
   console.log(`proxy: ${data}`);
 });
 
-proxy.stdout.on("exit", (code, signal) => {
+proxy.on("exit", (code, signal) => {
   console.log(`proxy returned (return code: ${code}, signal: ${signal})`);
 });
-
-// TODO this should be a flag, and we should have
-// npm run test # run the tests
-// npm run start # start the proxy
-if (true) {
-  /*
-   * Start the tests
-   */
-  const wdio = child_process.spawn("npm", ["run", "wdio"], {
-    env: { ...process.env, II_DAPP_URL: II_DAPP_URL },
-  });
-
-  wdio.stdout.on("data", (data) => {
-    console.log(`wdio: ${data}`);
-  });
-
-  wdio.stderr.on("data", (data) => {
-    console.log(`wdio: ${data}`);
-  });
-
-  wdio.on('exit', (code, signal) => {
-    console.log("Killing proxy");
-    proxy.kill();
-    if (code > 0) {
-      // if code is set and non-zero, bubble up error from wdio tests
-      throw new Error(`End-to-end tests returned with ${code}`);
-    } else if (signal) {
-      // otherwise, if the child was killed externally
-      throw new Error(`End-to-end tests killed with signal ${signal}`);
-    } else {
-      console.log('End-to-end tests finished successfully');
-    }
-  })
-}
